Video Processing/requirements.txt
fastapi
uvicorn
torch
transformers
opencv-python
pillow
pytesseract
python-multipart
scikit-learn
fastapi[all]
tqdm
numpy
